Only export one representation of linearization tokens
Move `Linearization` to an internal module.

Add note that I've added too much laziness to logging.
